Rename UTXO plugin directory to engine.
